make Mat(std::vector<_Tp>) constructor explicit to avoid unintended conversion of std::vector to cv::Mat fixes SF #2936166
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@2597 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/tests/cv/src/detectors_test.cpp
@@ -200,7 +200,7 @@ template <class T> bool CV_DetectorsTest::testDedector(const Mat& img, const T& 
                 ++foundCounter;
 
         }
-        if (foundCounter == case_num)
+        if (foundCounter == (int)case_num)
             result.push_back(okp);
     }
